refactor: rename filename squashedDict to squashedDictManager and move to hints folder
