+++ src/hints/squashedDictManager.ts
